Make heading levels descend step-by-step
We use h1 in the header, which is fine.

We use h3 in most parts of the body, which is fine.

However, we use h1 in the base layout for each page, which comes between
the header and the majority of the body, and that causes us to skip a
heading level, which can apparently cause issues for some accessibility
tools.
Switching it to h2 should set this right. If we feel that those headings
are in need of a bit more 'pop' then we can always adjust the style of h2.